+++ relay-cogs/src/lib.rs
@@ -187,8 +187,6 @@ pub enum AppFeature {
     MetricsTransactions,
     /// Metrics in the sessions namespace.
     MetricsSessions,
-    /// Metrics in the custom namespace.
-    MetricsCustom,
     /// Metrics in the unsupported namespace.
     ///
     /// This is usually not emitted, since metrics in the unsupported
@@ -220,7 +218,6 @@ impl AppFeature {
             Self::MetricsSpans => "metrics_spans",
             Self::MetricsTransactions => "metrics_transactions",
             Self::MetricsSessions => "metrics_sessions",
-            Self::MetricsCustom => "metrics_custom",
             Self::MetricsUnsupported => "metrics_unsupported",
             Self::Profiles => "profiles",
             Self::TraceMetrics => "trace_metrics",